backend/requirements.txt
flask
flask-cors
werkzeug
python-dotenv
numpy<2.0
opencv-python
matplotlib
nibabel
imageio
torch
torchvision
transformers
safetensors
Pillow
tensorflow
scikit-image
SQLAlchemy
psycopg2-binary

